get chromedriver version based on chrome version
Google only maintains the latest stable image of chrome, so we cannot get a specific version of google chrome. Instead, we can use the code in this PR to query the google chrome version and download the correct chromedriver (currently pulls 102.0.5005.61).  Source of this is from https://gerg.dev/2021/06/making-chromedriver-and-chrome-versions-match-in-a-docker-image/